# fastapi_backend/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "北京健康评估系统API",
    "description": "北京健康评估系统的后端API接口",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/models/": {
      "post": {
        "tags": ["模型管理"],
        "summary": "上传模型文件",
        "operationId": "create_model",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "model_type": {
                    "type": "integer",
                    "description": "模型类型：0=应激, 1=抑郁, 2=焦虑, 3=社交孤立"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": ["model_type", "file"]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "模型上传成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Model"
                }
              }
            }
          },
          "400": {
            "description": "无效的模型类型"
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          }
        }
      },
      "get": {
        "tags": ["模型管理"],
        "summary": "获取模型列表",
        "operationId": "read_models",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取模型列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Model"
                  }
                }
              }
            }
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          }
        }
      }
    },
    "/api/models/{model_id}": {
      "get": {
        "tags": ["模型管理"],
        "summary": "获取特定模型",
        "operationId": "read_model",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取模型",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Model"
                }
              }
            }
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "模型不存在"
          }
        }
      },
      "delete": {
        "tags": ["模型管理"],
        "summary": "删除模型",
        "operationId": "delete_model",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "model_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "成功删除模型"
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "模型不存在"
          }
        }
      }
    },
...
    "/api/parameters/": {
      "post": {
        "tags": ["参数管理"],
        "summary": "创建新系统参数",
        "operationId": "create_parameter",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParameterCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "参数创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Parameter"
                }
              }
            }
          },
          "400": {
            "description": "参数名已存在"
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          }
        }
      },
      "get": {
        "tags": ["参数管理"],
        "summary": "获取系统参数列表",
        "operationId": "read_parameters",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取参数列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Parameter"
                  }
                }
              }
            }
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          }
        }
      }
    },
    "/api/parameters/{param_id}": {
      "get": {
        "tags": ["参数管理"],
        "summary": "获取特定系统参数",
        "operationId": "read_parameter",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "param_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取参数",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Parameter"
                }
              }
            }
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "参数不存在"
          }
        }
      },
      "put": {
        "tags": ["参数管理"],
        "summary": "更新系统参数",
        "operationId": "update_parameter",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "param_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ParameterUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功更新参数",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Parameter"
                }
              }
            }
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "参数不存在"
          }
        }
      },
      "delete": {
        "tags": ["参数管理"],
        "summary": "删除系统参数",
        "operationId": "delete_parameter",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "param_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "成功删除参数"
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "参数不存在"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Model": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "model_type": {
            "type": "integer",
            "description": "模型类型：0=应激, 1=抑郁, 2=焦虑, 3=社交孤立"
          },
          "model_path": {
            "type": "string"
          },
          "upload_time": {
            "type": "string",
            "format": "date-time"
          },
          "upload_user": {
            "type": "string"
          },
          "is_active": {
            "type": "boolean"
          }
        }
      },
...
      "Parameter": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "ParameterCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": ["name", "value"]
      },
      "ParameterUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
} 